pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "localcull"
version = "0.6.1"
description = "AI-assisted local photo culling for RAW and JPEG workflows"
authors = [{name = "Arryn Panagos"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Elastic License 2.0 (ELv2)"}

dependencies = [
    "numpy>=1.26",
    "scipy>=1.11",
    "Pillow>=10.0",
    "opencv-python>=4.8",
    "tqdm>=4.66",
    "torch>=2.1",
    "pyiqa>=0.1.10",
    "mediapipe>=0.10",
    "PyExifTool>=0.5",
    "psutil>=5.9",
    "scikit-learn>=1.3",
    "mlx>=0.12",
    "mlx-image>=0.1.7",
]

[project.optional-dependencies]
mlx-vlm = ["mlx-vlm>=0.1"]
raw = ["rawpy>=0.20"]
dev = ["pytest>=7.0", "pytest-timeout>=2.2"]

[project.scripts]
localcull = "localcull.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "fixtures: requires test fixture CR3 files",
]